Split components to own files
